+++ mobile/android/geckoview/CHANGELOG.md
@@ -2,7 +2,7 @@
 - Moved `CompositorController`, `DynamicToolbarAnimator`,
   `OverscrollEdgeEffect`, `PanZoomController` from `org.mozilla.gecko.gfx` to
   `org.mozilla.geckoview`
-- Added `@UiThread`, `@AnyThread` annotations to several APIs
+- Added `@UiThread`, `@AnyThread` annotations to all APIs
 - Changed `GeckoRuntime#getLocale` to `GeckoRuntime#getLocales` and related APIs.
 - Merged `org.mozilla.gecko.gfx.LayerSession` into `GeckoSession`
 - Added `GeckoSession.MediaDelegate` and `MediaElement`. This allow monitoring
@@ -25,8 +25,9 @@
   `ContentDelegate#onContextMenu`. Extended information includes the element's
   title and alt attributes.
 - Changed `ContentDelegate.ContextElement` TYPE_ constants to public access.
-  Changed `ContentDelegate.ContextElement` to non-final class.
+- Changed `ContentDelegate.ContextElement`, `GeckoSession.FinderResult` to
+  non-final class.
 - Update `CrashReporter.sendCrashReport()` to return the crash ID as a
   GeckoResult<String>.
 
-[api-version]: d6c40ef4886c7818a446fb7b7a1cdbde706025ab
+[api-version]: bb945ae930ebf055f11821c685a6691faa7e5a3a